requirements.txt
# Core dependencies
openai>=1.12.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
supabase>=2.3.5

# Web scraping dependencies
playwright>=1.42.0
beautifulsoup4>=4.12.3
requests>=2.31.0

# Database migrations
alembic>=1.13.1
SQLAlchemy>=2.0.27 